refactor with parcel build
