+++ .vscode/launch.json
@@ -4,14 +4,12 @@
     // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
     "version": "0.2.0",
     "configurations": [
-
         {
             "type": "chrome",
             "request": "launch",
             "name": "Launch Chrome",
-            "url": "http://localhost:8080",
-            "preLaunchTask": "server",
-            "webRoot": "${workspaceFolder}"
+            "url": "http://mars.a.net:1234",
+            "webRoot": "${workspaceFolder}/app"
         }
     ]
 }
