+++ app/index.html
@@ -6,7 +6,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <title>Math World</title>
     <link rel='shortcut icon' type='image/x-icon' href='favicon.png'>
-    <script type="module" src="./index.js"></script>
     <style>
         * {
             font-family: consolas, monospace;
@@ -15,10 +14,10 @@
         body {
             margin: 0;
         }
-
     </style>
 </head>
 
 <body></body>
 
 </html>
+<script src="index.tsx"></script>
