+++ app/pages/math-single.js
