+++ tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "allowJs": true,
+    "target": "es6",
+    "module": "es6",
+    "moduleResolution": "node",
+    "noImplicitAny": false,
+    "removeComments": false,
+    "sourceMap": true,
+    "outDir": "./dist",
+    "jsx": "react",
+    "jsxFactory": "h",
+    "noEmit": true
+  },
+  "files": [
+    "./app/index.tsx"
+  ]
+}
